Mudancas na classe calculo

+++ Projeto_WindowsForms/Projeto_WindowsForms/Controle/Calculo.cs
@@ -3,19 +3,19 @@
     public class Calculo
     {
         // Propriedades
-        public double Inss { get; private set; }
-        public double Irpf { get; private set; }
-        public double Salario { get; private set; }
-        public double VencimentosTotais { get; private set; }
-        public double DescontosTotais { get; private set; }
-        public double ValorLiquido { get; private set; }
-        public double HorasExtras { get; private set; }
-        public double SalarioHora { get; private set; }
-        public double ResultadoHorasExtras { get; private set; }
+        public decimal Inss { get; private set; }
+        public decimal Irpf { get; private set; }
+        public decimal Salario { get; private set; }
+        public decimal VencimentosTotais { get; private set; }
+        public decimal DescontosTotais { get; private set; }
+        public decimal ValorLiquido { get; private set; }
+        public decimal HorasExtras { get; private set; }
+        public decimal SalarioHora { get; private set; }
+        public decimal ResultadoHorasExtras { get; private set; }
 
 
         // Métodos
-        public void CalcularSalarioLiquido(double salarioBase, double horasExtras)
+        public void CalcularSalarioLiquido(decimal salarioBase, int horasExtras)
         {
             ResultadoHorasExtras = CalcularHorasExtras(salarioBase, horasExtras);
             Inss = CalcularINSS(salarioBase);
@@ -30,23 +30,23 @@
             Salario = salarioBase;
         }
 
-        private double CalcularINSS(double salarioBase)
+        private decimal CalcularINSS(decimal salarioBase)
         {
-            Inss = salarioBase * 0.1; // 10% de desconto fictício
+            Inss = salarioBase * 0.1m; // 10% de desconto fictício
             Inss = Math.Round(Inss, 2);
             return Inss;
         }
 
-        private double CalcularIRPF(double salarioBase, double horasExtras)
+        private decimal CalcularIRPF(decimal salarioBase, decimal horasExtras)
         {
-            Irpf = (salarioBase + horasExtras) * 0.05; // 5% de desconto fictício
+            Irpf = (salarioBase + horasExtras) * 0.05m; // 5% de desconto fictício
             Irpf = Math.Round(Irpf, 2);
             return Irpf;
         }
-        private double CalcularHorasExtras(double salariobase, double horasExtras)
+        private decimal CalcularHorasExtras(decimal salariobase, decimal horasExtras)
         {
             SalarioHora = salariobase / 30 / 8;
-            HorasExtras = SalarioHora + SalarioHora * 0.5;
+            HorasExtras = SalarioHora + SalarioHora * 0.5m;
             ResultadoHorasExtras = HorasExtras * horasExtras;
             ResultadoHorasExtras = Math.Round(ResultadoHorasExtras, 2);
             return ResultadoHorasExtras;
