+++ Projeto_WindowsForms/Projeto_WindowsForms/Modelo/FolhaPagamento.cs
@@ -2,13 +2,13 @@
 {
     public class FolhaPagamento
     {
-        public double Inss { get; set; }
-        public double Irpf { get; set; }
-        public double VencimentosTotais { get; set; }
-        public double DescontosTotais { get; set; }
-        public double ValorLiquido { get; set; }
-        public double HorasExtras { get; set; }
-        public double SalarioHora { get; set; }
+        public decimal Inss { get; set; }
+        public decimal Irpf { get; set; }
+        public decimal VencimentosTotais { get; set; }
+        public decimal DescontosTotais { get; set; }
+        public decimal ValorLiquido { get; set; }
+        public decimal HorasExtras { get; set; }
+        public decimal SalarioHora { get; set; }
         public Colaborador Colaborador { get; set; }
     }
 }
